test: add an istanbul bootstrap script
When the test script is loaded it is already too late, and we are missing a lot
of information for coverage. This makes sure we get coverage for everything
(except bootstrap itself).
