+++ lib/bootstrap-local.js
@@ -13,6 +13,12 @@ const path = require('path');
 const ts = require('typescript');
 
 
+let _istanbulRequireHook = null;
+if (process.env['CODE_COVERAGE'] || process.argv.indexOf('--code-coverage') !== -1) {
+  _istanbulRequireHook = require('./istanbul-local').istanbulRequireHook;
+}
+
+
 // Check if we need to profile this CLI run.
 let profiler = null;
 if (process.env['DEVKIT_PROFILING']) {
@@ -45,8 +51,6 @@ Error.stackTraceLimit = Infinity;
 global._DevKitIsLocal = true;
 global._DevKitRoot = path.resolve(__dirname, '..');
 
-global._DevKitRequireHook = null;
-
 
 const compilerOptions = ts.readConfigFile(path.join(__dirname, '../tsconfig.json'), p => {
   return fs.readFileSync(p, 'utf-8');
@@ -71,8 +75,8 @@ require.extensions['.ts'] = function (m, filename) {
   try {
     let result = ts.transpile(source, compilerOptions['compilerOptions'], filename);
 
-    if (global._DevKitRequireHook) {
-      result = global._DevKitRequireHook(result, filename);
+    if (_istanbulRequireHook) {
+      result = _istanbulRequireHook(result, filename);
     }
 
     // Send it to node to execute.
